Move parts of ChessGame to ChessApplication class
- ChessApplication 클래스 생성
- main 메서드를 ChessApplication으로 이동
- ChessGame 클래스에 move 메서드 추가
